std::formatter<tds::datetime>::parse: fix clang warning

+++ src/tdscpp.h
@@ -2527,7 +2527,7 @@ struct std::formatter<tds::datetime> {
 
         if (it != ctx.end()) {
             if (*it >= '0' && *it <= '7') {
-                len = *it - '0';
+                len = (unsigned int)(*it - '0');
                 it++;
             } else if (*it == '{') {
                 it++;
